правка компоновки блока Filter
