+++ src/contexts/FiltersContext.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const FiltersContext = createContext();
