+++ src/css/order-form/order-form.css
@@ -6,8 +6,7 @@
   width: 30rem;
   height: 49rem;
   background: var(--color-style-white);
-  box-shadow: 0 0 4rem rgba(0, 0, 0, 0.2), 0 0 1rem rgba(0, 0, 0, 0.15),
-    0 0 0.25rem rgba(0, 0, 0, 0.1);
+  box-shadow: 0 0 4rem rgba(0, 0, 0, 0.2), 0 0 1rem rgba(0, 0, 0, 0.15), 0 0 0.25rem rgba(0, 0, 0, 0.1);
 }
 
 .order-form__card-orders {
@@ -141,14 +140,14 @@
   width: 100%;
 }
 
-.order-form__order-status .multipleStatus {
+.order-form__order-status ._ {
   display: none;
   position: absolute;
   right: 0;
   bottom: 2rem;
 }
 
-.order-form__order-status .input:hover + .multipleStatus {
+.order-form__order-status .input:hover + ._ {
   display: flex;
 }
