+++ src/modules/Filter/Filter.module.css
@@ -1,0 +1,7 @@
+._ {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 1rem;
+  width: 100%;
+}
