+++ src/modules/Filter/FilterSearchPanel/FilterSearchPanel.module.css
@@ -1,0 +1,18 @@
+._ {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  height: 2rem;
+}
+
+.searchContainer {
+  display: flex;
+  align-items: flex-start;
+  gap: 0.5rem;
+  width: 38rem;
+}
+
+.searchContainer .searchbar {
+  flex-grow: 1;
+}
