+++ src/modules/Filter/FilterSelectorsPanel/FilterOrderStatus/FilterOrderStatus.module.css
@@ -1,0 +1,23 @@
+._ {
+  top: 3.7rem;
+  right: 0;
+  align-items: flex-start;
+  padding: 0.5rem 0;
+}
+
+.checkbox {
+  align-self: stretch;
+  padding: 0.5rem 0 0.5rem 1rem;
+}
+
+.multiple {
+  height: 1rem;
+  width: 1rem;
+  transform: rotate(90deg);
+  stroke: var(--color-style-light-blue);
+  fill: var(--color-style-light-blue);
+}
+
+.multiple:hover {
+  transform: rotate(0deg);
+}
