+++ src/modules/Filter/FilterSelectorsPanel/FiltersDateRegistration/FiltersDateRegistration.module.css
@@ -1,0 +1,6 @@
+._ {
+  display: flex;
+  align-items: flex-end;
+  gap: 0.5rem;
+  width: 18rem;
+}
