--- src/modules/Filters/Filters.module.css
@@ -1,56 +1,0 @@
-._ {
-  display: flex;
-  align-items: flex-end;
-  padding: 1rem;
-  gap: 1rem;
-  width: 100%;
-  height: 5.5rem;
-  background-color: var(--color-style-white-blue);
-  border-radius: 0.5rem;
-}
-
-.registration {
-  display: flex;
-  align-items: flex-end;
-  gap: 0.5rem;
-  width: 18rem;
-}
-
-.registration .input {
-  flex-basis: 50%;
-}
-
-.amount {
-  display: flex;
-  align-items: flex-end;
-  gap: 0.5rem;
-  width: 16rem;
-}
-
-.amount .input {
-  flex-basis: 50%;
-}
-
-.multipleStatus {
-  top: 3.7rem;
-  right: 0;
-  align-items: flex-start;
-  padding: 0.5rem 0;
-}
-
-.checkbox {
-  align-self: stretch;
-  padding: 0.5rem 0 0.5rem 1rem;
-}
-
-.multiple {
-  height: 1rem;
-  width: 1rem;
-  transform: rotate(90deg);
-  stroke: var(--color-style-light-blue);
-  fill: var(--color-style-light-blue);
-}
-
-.multiple:hover {
-  transform: rotate(0deg);
-}
