--- src/modules/Search/Search.module.css
@@ -1,26 +1,0 @@
-._ {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 1rem;
-  width: 100%;
-}
-
-.topPanel {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  height: 2rem;
-}
-
-.searchContainer {
-  display: flex;
-  align-items: flex-start;
-  gap: 0.5rem;
-  width: 38rem;
-}
-
-.searchContainer .searchbar {
-  flex-grow: 1;
-}
